+++ www/commands.js
@@ -15,7 +15,7 @@ const xmlEscapeAttribute = (value) => {
 	}
 	// Temporarily DO NOT encode the & character in XML attributes
 	return String(value)
-		//.replace(/&/g, '&amp;')
+		.replace(/&/g, '&amp;')
 		.replace(/"/g, '&quot;')
 		.replace(/</g, '&lt;')
 		.replace(/>/g, '&gt;')